git-svn-id: svn://danse.us/AbInitio/espresso/qecalc@754 4efde5e4-ebd3-4001-8375-eeb568285340

+++ qecalc/qetask/qetask.py
@@ -35,7 +35,8 @@ class QETask(object):
             raise Exception("Task " + self.name + " crashed: check your settings" + "Command string:" + self.cmdStr)
 
     def _run(self):
-        os.remove('CRASH')
+        if os.path.exists('CRASH'):
+            os.remove('CRASH')
         if self.setting.paraPrefix != '' and self.setting.paraPrefix in self.cmdStr:
             if self.setting.useTorque:
                 self.torque.serial(self.cmdStr)
